+++ casual/casual-event-server/src/main/java/se/laz/casual/event/server/handlers/EventMessageEncoder.java
@@ -4,7 +4,7 @@
  * This software is licensed under the MIT license, https://opensource.org/licenses/MIT
  */
 
-package se.laz.casual.event.server;
+package se.laz.casual.event.server.handlers;
 
 import io.netty.buffer.ByteBuf;
 import io.netty.channel.ChannelHandlerContext;
