+++ casual/casual-event-server/src/main/java/se/laz/casual/event/server/handlers/ExceptionHandler.java
@@ -4,7 +4,7 @@
  * This software is licensed under the MIT license, https://opensource.org/licenses/MIT
  */
 
-package se.laz.casual.event.server;
+package se.laz.casual.event.server.handlers;
 
 import io.netty.channel.ChannelHandler;
 import io.netty.channel.ChannelHandlerContext;
