+++ casual/casual-event-server/src/main/java/se/laz/casual/event/server/handlers/FromJSONLogonDecoder.java
@@ -1,4 +1,4 @@
-package se.laz.casual.event.server;
+package se.laz.casual.event.server.handlers;
 
 import io.netty.buffer.ByteBuf;
 import io.netty.channel.ChannelHandlerContext;
@@ -6,6 +6,8 @@ import io.netty.channel.SimpleChannelInboundHandler;
 import io.netty.channel.group.ChannelGroup;
 import io.netty.util.CharsetUtil;
 import se.laz.casual.api.external.json.JsonProviderFactory;
+import se.laz.casual.event.server.messages.LogonRequestMessage;
+import se.laz.casual.event.server.messages.LogonRequestTypeAdapter;
 
 import java.util.Objects;
 import java.util.logging.Logger;
@@ -34,6 +36,7 @@ public class FromJSONLogonDecoder extends SimpleChannelInboundHandler<Object>
         log.info(() -> "msg json: " + json);
         LogonRequestMessage requestMessage = JsonProviderFactory.getJsonProvider().fromJson(json, LogonRequestMessage.class, LogonRequestTypeAdapter.of());
         connectedClients.add(ctx.channel());
+        ctx.fireChannelRead(requestMessage);
         log.info(() -> "client logged on: " + requestMessage);
     }
 }
