+++ casual/casual-event-server/src/main/java/se/laz/casual/event/server/messages/LogonRequest.java
@@ -1,4 +1,4 @@
-package se.laz.casual.event.server;
+package se.laz.casual.event.server.messages;
 
 import se.laz.casual.api.CasualRuntimeException;
